requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Google AI
google-generativeai>=0.8.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Vector store (for knowledge base)
chromadb>=0.4.0

# API utilities
python-multipart>=0.0.6
aiofiles>=23.2.1

# HTTP client
httpx>=0.25.0

# Telegram Bot (optional)
python-telegram-bot>=20.7

# Voice (optional)
# google-cloud-speech>=2.23.0
# google-cloud-texttospeech>=2.14.0
